requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.18.0
idna==3.10
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
tldextract==5.2.0
urllib3==2.4.0